scraper/requirements.txt
scrapy==2.11.0
beautifulsoup4==4.12.2
lxml==5.1.0
python-dotenv==1.0.0
fake-useragent==1.4.0
itemadapter==0.8.0
